Updated Simulator.py to handle new testbed I/O specification.
This update was done assuming that the testbed can handle receiving
payoffs for a strict subset of the players (because some players
default, they have no payoff). This is not currently true, but might be
soon. I might also consider a DPR-friendly version that re-simulates if
the only player following a given strategy defaults.